# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "FKMed",
    "description" : "FKMed — health plan beneficiary portal API",
    "version" : "0.12.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/reimbursements" : {
      "get" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "history",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReimbursementHistoryItem"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "submit",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "request" : {
                    "$ref" : "#/components/schemas/ReimbursementSubmitRequest"
                  },
                  "documents" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "request" ]
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementSubmissionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/reimbursements/{id}/pendency-documents" : {
      "post" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "resolvePendency",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "request" : {
                    "$ref" : "#/components/schemas/ReimbursementPendencyDocumentsRequest"
                  },
                  "documents" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "request" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementActionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/reimbursements/{id}/bank-correction" : {
      "post" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "correctBank",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReimbursementBankCorrectionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementActionResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/recovery/request" : {
      "post" : {
        "tags" : [ "recovery-controller" ],
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRecoveryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        }
      }
    },
...
    "/api/auth/first-access/complete" : {
      "post" : {
        "tags" : [ "first-access-controller" ],
        "operationId" : "complete",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompleteFirstAccessRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        }
      }
    },
...
    "/api/tokens/current" : {
      "get" : {
        "tags" : [ "token-controller" ],
        "operationId" : "current",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reimbursements/{id}" : {
      "get" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementDetailView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reimbursements/statement" : {
      "get" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "statement",
        "parameters" : [ {
          "name" : "beneficiaryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReimbursementStatementView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reimbursements/documentation-guide" : {
      "get" : {
        "tags" : [ "reimbursement-controller" ],
        "operationId" : "documentationGuide",
        "parameters" : [ {
          "name" : "type",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentationGuideView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/legal-documents/{type}" : {
      "get" : {
        "tags" : [ "legal-document-controller" ],
        "operationId" : "document",
        "parameters" : [ {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LegalDocumentView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TokenView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ReimbursementDocumentRequest" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string",
            "enum" : [ "RECEIPT", "MEDICAL_ORDER", "COMPLEMENTARY", "BUDGET" ]
          },
          "fileName" : {
            "type" : "string"
          }
        },
        "required" : [ "category" ]
      },
      "ReimbursementSessionRequest" : {
        "type" : "object",
        "properties" : {
          "sessionDate" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          }
        },
        "required" : [ "amount", "sessionDate" ]
      },
      "ReimbursementSubmitRequest" : {
        "type" : "object",
        "properties" : {
          "beneficiaryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "expenseTypeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "careDate" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          },
          "sessions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReimbursementSessionRequest"
            }
          },
          "providerName" : {
            "type" : "string",
            "minLength" : 1
          },
          "providerCouncilCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "providerCouncilNumber" : {
            "type" : "string",
            "minLength" : 1
          },
          "providerCouncilUf" : {
            "type" : "string",
            "minLength" : 1
          },
          "providerDocument" : {
            "type" : "string",
            "minLength" : 1
          },
          "providerSpecialty" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAgency" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAccount" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAccountDigit" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAccountType" : {
            "type" : "string",
            "enum" : [ "CORRENTE", "POUPANCA" ]
          },
          "acceptedTermVersion" : {
            "type" : "string",
            "minLength" : 1
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReimbursementDocumentRequest"
            }
          }
        },
        "required" : [ "acceptedTermVersion", "amount", "bankAccount", "bankAccountDigit", "bankAccountType", "bankAgency", "bankCode", "beneficiaryId", "careDate", "expenseTypeCode", "providerCouncilCode", "providerCouncilNumber", "providerCouncilUf", "providerDocument", "providerName", "providerSpecialty" ]
      },
      "ReimbursementSubmissionResult" : {
        "type" : "object",
        "properties" : {
          "protocol" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "expectedPaymentDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ReimbursementPendencyDocumentsRequest" : {
        "type" : "object",
        "properties" : {
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReimbursementDocumentRequest"
            }
          }
        }
      },
      "ReimbursementActionResult" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "expectedPaymentDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ReimbursementBankCorrectionRequest" : {
        "type" : "object",
        "properties" : {
          "bankCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAgency" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAccount" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAccountDigit" : {
            "type" : "string",
            "minLength" : 1
          },
          "bankAccountType" : {
            "type" : "string",
            "enum" : [ "CORRENTE", "POUPANCA" ]
          }
        },
        "required" : [ "bankAccount", "bankAccountDigit", "bankAccountType", "bankAgency", "bankCode" ]
      },
...
      "PasswordRecoveryRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          }
        },
        "required" : [ "email" ]
      },
...
      "CompleteFirstAccessRequest" : {
        "type" : "object",
        "properties" : {
          "registrationToken" : {
            "type" : "string",
            "minLength" : 1
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 160,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          },
          "acceptedTerms" : {
            "type" : "boolean"
          },
          "acceptedPrivacy" : {
            "type" : "boolean"
          }
        },
        "required" : [ "email", "password", "registrationToken" ]
      },
...
      "ReimbursementHistoryItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "expenseType" : {
            "type" : "string"
          },
          "beneficiary" : {
            "type" : "string"
          },
          "requestedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "amountRequested" : {
            "type" : "number"
          },
          "amountReimbursed" : {
            "type" : "number"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          }
        }
      },
      "BankView" : {
        "type" : "object",
        "properties" : {
          "bankCode" : {
            "type" : "string"
          },
          "agency" : {
            "type" : "string"
          },
          "account" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CORRENTE", "POUPANCA" ]
          }
        }
      },
      "DocumentView" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "type" : "string",
            "enum" : [ "RECEIPT", "MEDICAL_ORDER", "COMPLEMENTARY", "BUDGET" ]
          },
          "fileName" : {
            "type" : "string"
          },
          "fileSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "uploadedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "GlosaView" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
      "PendencyView" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "deadlineAt" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ProviderView" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "councilCode" : {
            "type" : "string"
          },
          "councilNumber" : {
            "type" : "string"
          },
          "councilUf" : {
            "type" : "string"
          },
          "specialty" : {
            "type" : "string"
          }
        }
      },
      "ReimbursementDetailView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "expenseType" : {
            "type" : "string"
          },
          "beneficiary" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "careDate" : {
            "type" : "string",
            "format" : "date"
          },
          "requestedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expectedPaymentDate" : {
            "type" : "string",
            "format" : "date"
          },
          "amountRequested" : {
            "type" : "number"
          },
          "amountReimbursed" : {
            "type" : "number"
          },
          "glosa" : {
            "$ref" : "#/components/schemas/GlosaView"
          },
          "denialReason" : {
            "type" : "string"
          },
          "pendency" : {
            "$ref" : "#/components/schemas/PendencyView"
          },
          "bank" : {
            "$ref" : "#/components/schemas/BankView"
          },
          "provider" : {
            "$ref" : "#/components/schemas/ProviderView"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DocumentView"
            }
          },
          "timeline" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimelineView"
            }
          },
          "regulatoryNote" : {
            "type" : "string"
          }
        }
      },
      "TimelineView" : {
        "type" : "object",
        "properties" : {
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANALISE", "PROCESSAMENTO", "PENDENTE_DOCUMENTACAO", "APROVADO", "PAGO", "PAGAMENTO_NAO_EFETUADO", "NEGADO", "CANCELADO" ]
          },
          "description" : {
            "type" : "string"
          }
        }
      },
...
      "ReimbursementStatementView" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/StatementItem"
            }
          },
          "total" : {
            "type" : "number"
          }
        }
      },
      "StatementItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "protocol" : {
            "type" : "string"
          },
          "beneficiary" : {
            "type" : "string"
          },
          "paidAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "amountPaid" : {
            "type" : "number"
          }
        }
      },
...
      "DocumentationGuideView" : {
        "type" : "object",
        "properties" : {
          "expenseType" : {
            "type" : "string"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "LegalDocumentView" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "publishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "body" : {
            "type" : "string"
          },
          "acceptedByMe" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}